+++ src/test/e2e/services/client-metrics-service.e2e.test.js
@@ -19,26 +19,28 @@ beforeAll(async () => {
 });
 
 afterAll(async () => {
-    await db.destroy();
+    if (db) {
+        await db.destroy();
+    }
 });
 test('Apps registered should be announced', async () => {
     expect.assertions(3);
     const clientRegistration = {
         appName: faker.internet.domainName(),
-        instanceId: faker.random.uuid(),
+        instanceId: faker.datatype.uuid(),
         strategies: ['default'],
         started: Date.now(),
-        interval: faker.random.number(),
+        interval: faker.datatype.number(),
         icon: '',
         description: faker.company.catchPhrase(),
         color: faker.internet.color(),
     };
     const differentClient = {
         appName: faker.lorem.slug(2),
-        instanceId: faker.random.uuid(),
+        instanceId: faker.datatype.uuid(),
         strategies: ['default'],
         started: Date.now(),
-        interval: faker.random.number(),
+        interval: faker.datatype.number(),
         icon: '',
         description: faker.company.catchPhrase(),
         color: faker.internet.color(),
